src/gen-data/python/requirements.txt
GitPython==2.1.11
PyGithub==1.43.3